Create modules to add make and add projects
This change creates two more modules that creates
new projects and adds those projects to a projects
collection object, similar in structure to how task
items will be organized in task maker and task
adder modules. The button/update state will
create a new project, add it to the projects
collections, then create a new task item, and add
that inside the project.

This temporary code will work for adding a single
task item to a single new project.