Change createItem function
This change made two new functions specific to
projects and todos instead of using a shared
create item function which loops through properties.
The reason for this change is because I need to
layout the todo properties in its container in a
certain way instead of just looping and placing
them automatically with flexbox.