Fix layout for todos
Changed version of layout for todos.
All todos will have checkbox. If no title exist,
default "Untitled" is assigned. Ellipsis form on
long titles. Ellipsis also form if todos contain
a description, which can be expanded to view the
description. Position of due date will be the same
regardless of due date or not.