Remove projectId from properties
Remove projectId from todo properties as it was being
used in a roundabout way to retrieve project title by
looking 2 layers up, calling Project Manager to get
the project using the id, and then the project's
title. Now that a select element is added to the HTML
with the project titles as options, the project
title can directly be passed into the form object
on todo submission/creation.