Remove reliance on data attributes
Data attributes on project form and todo form was
previously used to set the context on which project
or todo was being operated on.

This change not only added a dropdown that can select
the context manually, later implemented in global
view such as the filtered tabs, but also added
properties on project manager, which already keeps
track of global state, to keep track on current
project/todo context.