Update context flow across projects, todos, and filter
App starts in inbox filter context with no set
project, allowing users to select from a dropdown
which project their new todos can go into, with
default project being an option.

Creating a new project clears the filter (global view)
context, and sets the current project as the context.

Within the current project, adding a new todo
automatically prefills the select dropdown to the
in view project as the context. Editing a todo
within a project context also prefills the select
dropdown.

Switching to the filter tabs sets the context to
the in view filters, allowing users to once again
add todos to their desired list of projects.

Additionally, editing todos from this global view
filter context will edit "in place", as each todo
is agnostic of which project their belong to, therefore
the select dropdown will be prefilled with the project
they belong to.

Additionaly reset current project lines in form
submit handlers and dialog cancel event to ensure
that the context flow is rigid.