Replace project dialog form with inline input
Remove work flow for creating and editing projects.
Previously creating and editing had to go through
a dialog, even though the only required input is
the title of the project. Replaced that flow with
creating and editing inline without dialog.

Could reimplement dialog in the future if creating
or editing a project requires more inputs.