Extract form cancellation function
This change also checks for if user has actually
interacted with the project form via checking
for input's value before and after blur.

Previously, clicking to edit a project title but
not actually changing anything would still cause
the form the submit, and update the project
"in place" with duplicate data. This was inefficient
and also caused the side effect of rendering the
the todos for the clicked/edited project, which
was unintentional.

This change added a check for input's before and
after blur value to circumvent that, while also
extracted identical form cancellation function.