Add has submitted guard to submit and cancel form
Previously, was only in submit form, causing input
blur event to fire simultaneously with keydown
and submit events when the form is cancelled (with
esc) or the form is submitted with no change/blank.

Adding the hasSubmitted guard to cancelForm now
correctly stops other events from firing and
ensures only one event fires.