Update dropdown menu positioning
It is quite literally impossible with my current
setup to get the top left corner/area of the
dropdown menu, to align and overlap slightly with
the bottom right corner of the .project.

Doing top: 85%, left: 80% creates the position of
where I would like the dropdown to appear (top left
of dropdown overlapping slightly with bottom right
of project), but the problem is, a horizontal
scrollbar now appears in projects container.

It appears it's due to my constraints on overflow,
in order to maintain my hover scrollbar effect,
and a combination of flex containers with no fixed
widths.

This positioning is the best I can do for this
current setup without having to refactor my CSS.